requirements.txt
altgraph==0.17.4
black==25.1.0
click==8.1.8
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pefile==2024.8.26
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
setuptools==75.8.2
